requirements.txt
asgiref==3.5.0
Django==4.0.2
django-js-asset==2.0.0
django-mptt==0.13.4
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2021.5
